+++ spec/SubStatementSpec.php
@@ -14,12 +14,14 @@ namespace spec\Xabbuh\XApi\Model;
 use PhpSpec\ObjectBehavior;
 use Xabbuh\XApi\Model\Activity;
 use Xabbuh\XApi\Model\Agent;
+use Xabbuh\XApi\Model\Attachment;
 use Xabbuh\XApi\Model\Context;
 use Xabbuh\XApi\Model\ContextActivities;
 use Xabbuh\XApi\Model\Extensions;
 use Xabbuh\XApi\Model\Group;
 use Xabbuh\XApi\Model\InverseFunctionalIdentifier;
 use Xabbuh\XApi\Model\IRI;
+use Xabbuh\XApi\Model\IRL;
 use Xabbuh\XApi\Model\LanguageMap;
 use Xabbuh\XApi\Model\Result;
 use Xabbuh\XApi\Model\StatementId;
@@ -45,6 +47,41 @@ class SubStatementSpec extends ObjectBehavior
         $this->beConstructedWith($actor, $verb, $object);
 
         $this->shouldHaveType('Xabbuh\XApi\Model\Object');
+    }
+
+    function its_object_can_be_an_agent()
+    {
+        $actor = new Agent(InverseFunctionalIdentifier::withMbox(IRI::fromString('mailto:conformancetest@tincanapi.com')));
+        $verb = new Verb(IRI::fromString('http://tincanapi.com/conformancetest/verbid'), LanguageMap::create(array('en-US' => 'test')));
+        $object = new Agent(InverseFunctionalIdentifier::withOpenId('http://openid.tincanapi.com'));
+        $this->beConstructedWith($actor, $verb, $object);
+
+        $this->getObject()->shouldBeAnInstanceOf('Xabbuh\XApi\Model\Object');
+        $this->getObject()->shouldBe($object);
+    }
+
+    function it_does_not_equal_another_statement_with_different_timestamp()
+    {
+        $actor = new Agent(InverseFunctionalIdentifier::withMbox(IRI::fromString('mailto:conformancetest@tincanapi.com')));
+        $verb = new Verb(IRI::fromString('http://tincanapi.com/conformancetest/verbid'), LanguageMap::create(array('en-US' => 'test')));
+        $object = new Agent(InverseFunctionalIdentifier::withOpenId('http://openid.tincanapi.com'));
+        $this->beConstructedWith($actor, $verb, $object, null, null, new \DateTime('2014-07-23T12:34:02-05:00'));
+
+        $otherStatement = new SubStatement($actor, $verb, $object, null, null, new \DateTime('2015-07-23T12:34:02-05:00'));
+
+        $this->equals($otherStatement)->shouldBe(false);
+    }
+
+    function it_equals_another_statement_with_same_timestamp()
+    {
+        $actor = new Agent(InverseFunctionalIdentifier::withMbox(IRI::fromString('mailto:conformancetest@tincanapi.com')));
+        $verb = new Verb(IRI::fromString('http://tincanapi.com/conformancetest/verbid'), LanguageMap::create(array('en-US' => 'test')));
+        $object = new Agent(InverseFunctionalIdentifier::withOpenId('http://openid.tincanapi.com'));
+        $this->beConstructedWith($actor, $verb, $object, null, null, new \DateTime('2014-07-23T12:34:02-05:00'));
+
+        $otherStatement = new SubStatement($actor, $verb, $object, null, null, new \DateTime('2014-07-23T12:34:02-05:00'));
+
+        $this->equals($otherStatement)->shouldBe(true);
     }
 
     function it_is_different_from_another_sub_statement_if_contexts_differ()
@@ -129,6 +166,16 @@ class SubStatementSpec extends ObjectBehavior
         $subStatement->getResult()->shouldReturn($result);
     }
 
+    public function it_returns_a_new_instance_with_timestamp()
+    {
+        $timestamp = new \DateTime('2014-07-23T12:34:02-05:00');
+        $statement = $this->withTimestamp($timestamp);
+
+        $statement->shouldNotBe($this);
+        $statement->shouldBeAnInstanceOf('\Xabbuh\XApi\Model\SubStatement');
+        $statement->getTimestamp()->shouldReturn($timestamp);
+    }
+
     public function it_returns_a_new_instance_with_context()
     {
         $context = new Context();
@@ -137,5 +184,50 @@ class SubStatementSpec extends ObjectBehavior
         $subStatement->shouldNotBe($this);
         $subStatement->shouldBeAnInstanceOf('\Xabbuh\XApi\Model\SubStatement');
         $subStatement->getContext()->shouldReturn($context);
+    }
+
+    public function it_returns_a_new_instance_with_attachments()
+    {
+        $attachments = array(new Attachment(
+            IRI::fromString('http://id.tincanapi.com/attachment/supporting_media'),
+            'text/plain',
+            18,
+            'bd1a58265d96a3d1981710dab8b1e1ed04a8d7557ea53ab0cf7b44c04fd01545',
+            LanguageMap::create(array('en-US' => 'Text attachment')),
+            LanguageMap::create(array('en-US' => 'Text attachment description')),
+            IRL::fromString('http://tincanapi.com/conformancetest/attachment/fileUrlOnly')
+        ));
+        $statement = $this->withAttachments($attachments);
+
+        $statement->shouldNotBe($this);
+        $statement->shouldBeAnInstanceOf('\Xabbuh\XApi\Model\SubStatement');
+        $statement->getAttachments()->shouldReturn($attachments);
+    }
+
+    function it_ignores_array_keys_in_attachment_lists()
+    {
+        $textAttachment = new Attachment(
+            IRI::fromString('http://id.tincanapi.com/attachment/supporting_media'),
+            'text/plain',
+            18,
+            'bd1a58265d96a3d1981710dab8b1e1ed04a8d7557ea53ab0cf7b44c04fd01545',
+            LanguageMap::create(array('en-US' => 'Text attachment')),
+            LanguageMap::create(array('en-US' => 'Text attachment description')),
+            IRL::fromString('http://tincanapi.com/conformancetest/attachment/fileUrlOnly')
+        );
+        $attachments = array(1 => $textAttachment);
+
+        $actor = new Agent(InverseFunctionalIdentifier::withMbox(IRI::fromString('mailto:conformancetest@tincanapi.com')));
+        $verb = new Verb(IRI::fromString('http://tincanapi.com/conformancetest/verbid'), LanguageMap::create(array('en-US' => 'test')));
+        $object = new Activity(IRI::fromString('http://tincanapi.com/conformancetest/activityid'));
+        $this->beConstructedWith($actor, $verb, $object, null, null, null, $attachments);
+
+        $this->getAttachments()->shouldBeArray();
+        $this->getAttachments()->shouldHaveKeyWithValue(0, $textAttachment);
+
+        $statement = $this->withAttachments($attachments);
+
+        $statement->getAttachments()->shouldBeArray();
+        $statement->getAttachments()->shouldHaveKeyWithValue(0, $textAttachment);
     }
 }